Revert part of rebrand:
* Text domain needs to match WordPress.org plugin slug
* Sub directory should match parent directory
* Make sure admin area assets are loaded in Network admin

WP User Signups vs. WP Signups is confusing anyways, because what else is a sign-up? The fact that blogs are "sign-ups" too is just a WordPress quirk that does not necessarily need to be called out and made prominent by this plugin. So here we are, half-in and half-out, and I think it's probably still gonna be OK.